requirements.txt
ansible
boto3
docopt
PyGithub
